--- library/src/com/viewpagerindicator/BannerView.java
@@ -1,46 +1,0 @@
-package com.viewpagerindicator;
-
-import android.content.Context;
-import android.util.AttributeSet;
-import android.view.Gravity;
-import android.view.ViewGroup;
-import android.widget.FrameLayout;
-import android.widget.TextView;
-
-/**
- * Created By: andrewgrosner
- * Date: 10/18/13
- * Contributors:
- * Description:
- */
-public class BannerView extends FrameLayout{
-
-    public interface IBannerView{
-        public int getBannerWidth();
-    }
-
-    public BannerView(Context context) {
-        super(context);
-        setupUI(context);
-    }
-
-    public BannerView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-        setupUI(context);
-    }
-
-    public BannerView(Context context, AttributeSet attrs, int defStyle) {
-        super(context, attrs, defStyle);
-        setupUI(context);
-    }
-
-    protected void setupUI(Context context){
-        TextView title = new TextView(context);
-        title.setText("BannerPageIndicator");
-        title.setTextSize(75);
-        title.setGravity(Gravity.CENTER);
-
-        addView(title, new LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
-    }
-
-}
